+++ src/Parse.js
@@ -16,7 +16,7 @@ import InstallationController from './InstallationController';
 import * as ParseOp from './ParseOp';
 import RESTController from './RESTController';
 import MoralisWeb3 from './MoralisWeb3';
-const { checkForSdkUpdates } = require('./utils');
+const { checkForSdkUpdates, trackEvent, TrackingEventName, getSubdomain } = require('./utils');
 import { ethers } from 'ethers';
 import { filterConsole } from './filterConsole';
 
@@ -69,11 +69,31 @@ class Moralis extends MoralisWeb3 {
     this.Web3API.initialize({ serverUrl, apiKey, Moralis });
     this.SolanaAPI.initialize({ serverUrl, apiKey, Moralis });
     if (appId && serverUrl) {
+      trackOptions = {
+        subdomain: getSubdomain(serverUrl),
+        sdk_type: 'javascript',
+        sdk_version: process.env.NEXT_VERSION,
+        sdk_enviroment: process.env.PARSE_BUILD,
+        appId: appId,
+      };
       await this.initPlugins(plugins);
+    } else {
+      trackOptions = {
+        sdk_type: 'javascript',
+        sdk_version: process.env.NEXT_VERSION,
+        sdk_enviroment: process.env.PARSE_BUILD,
+      };
     }
 
     // Check if SDK is updated
     checkForSdkUpdates();
+
+    // Track start function call
+    trackEvent(
+      TrackingEventName.START_FUNCTION,
+      serverUrl ? getSubdomain(serverUrl) : null,
+      trackOptions
+    );
   }
 
   /**